CA-135774: Limit memory used when under memory pressure
Force tapdisk to process requests 1 at a time when under memory pressure to
limit the amount of memory used.  Always keep at least 1 request in the
bufcache for this purpose.

Use memory cgroups to listen to memory pressure notifications.  Use a timer and
a low water mark (set at 80%) to determine when to exit low memory mode.

Signed-off-by: Ross Lagerwall <ross.lagerwall@citrix.com>